remove discovery pkg
Currently unused, and functionality is largely duplicated
by the services/m365 package.
