+++ src/pkg/services/m365/m365_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/alcionai/clues"
+	"github.com/google/uuid"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
 	"github.com/microsoftgraph/msgraph-sdk-go/models/odataerrors"
 	"github.com/stretchr/testify/assert"
@@ -12,11 +13,13 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/alcionai/corso/src/internal/common/ptr"
-	"github.com/alcionai/corso/src/internal/m365/discovery"
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/tester"
+	"github.com/alcionai/corso/src/pkg/account"
+	"github.com/alcionai/corso/src/pkg/credentials"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/path"
+	"github.com/alcionai/corso/src/pkg/services/m365/api"
 )
 
 type M365IntegrationSuite struct {
@@ -183,13 +186,13 @@ func (m mockDGDD) GetDefaultDrive(context.Context, string) (models.Driveable, er
 func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 	table := []struct {
 		name      string
-		mock      func(context.Context) discovery.GetDefaultDriver
+		mock      func(context.Context) getDefaultDriver
 		expect    assert.BoolAssertionFunc
 		expectErr func(*testing.T, error)
 	}{
 		{
 			name: "ok",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				return mockDGDD{models.NewDrive(), nil}
 			},
 			expect: assert.True,
@@ -199,7 +202,7 @@ func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 		},
 		{
 			name: "mysite not found",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				odErr := odataerrors.NewODataError()
 				merr := odataerrors.NewMainError()
 				merr.SetCode(ptr.To("code"))
@@ -215,7 +218,7 @@ func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 		},
 		{
 			name: "mysite URL not found",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				odErr := odataerrors.NewODataError()
 				merr := odataerrors.NewMainError()
 				merr.SetCode(ptr.To("code"))
@@ -231,7 +234,7 @@ func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 		},
 		{
 			name: "no sharepoint license",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				odErr := odataerrors.NewODataError()
 				merr := odataerrors.NewMainError()
 				merr.SetCode(ptr.To("code"))
@@ -247,7 +250,7 @@ func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 		},
 		{
 			name: "user not found",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				odErr := odataerrors.NewODataError()
 				merr := odataerrors.NewMainError()
 				merr.SetCode(ptr.To(string(graph.RequestResourceNotFound)))
@@ -263,7 +266,7 @@ func (suite *m365UnitSuite) TestCheckUserHasDrives() {
 		},
 		{
 			name: "arbitrary error",
-			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+			mock: func(ctx context.Context) getDefaultDriver {
 				odErr := odataerrors.NewODataError()
 				merr := odataerrors.NewMainError()
 				merr.SetCode(ptr.To("code"))
@@ -360,6 +363,243 @@ func (suite *m365UnitSuite) TestGetAllSites() {
 
 			_, err := getAllSites(ctx, gas)
 			test.expectErr(t, err)
+		})
+	}
+}
+
+type DiscoveryIntgSuite struct {
+	tester.Suite
+	acct account.Account
+}
+
+func TestDiscoveryIntgSuite(t *testing.T) {
+	suite.Run(t, &DiscoveryIntgSuite{
+		Suite: tester.NewIntegrationSuite(
+			t,
+			[][]string{tester.M365AcctCredEnvs}),
+	})
+}
+
+func (suite *DiscoveryIntgSuite) SetupSuite() {
+	suite.acct = tester.NewM365Account(suite.T())
+}
+
+func (suite *DiscoveryIntgSuite) TestUsers() {
+	t := suite.T()
+
+	ctx, flush := tester.NewContext(t)
+	defer flush()
+
+	errs := fault.New(true)
+
+	users, err := Users(ctx, suite.acct, errs)
+	assert.NoError(t, err, clues.ToCore(err))
+
+	ferrs := errs.Errors()
+	assert.Nil(t, ferrs.Failure)
+	assert.Empty(t, ferrs.Recovered)
+	assert.NotEmpty(t, users)
+}
+
+func (suite *DiscoveryIntgSuite) TestUsers_InvalidCredentials() {
+	table := []struct {
+		name string
+		acct func(t *testing.T) account.Account
+	}{
+		{
+			name: "Invalid Credentials",
+			acct: func(t *testing.T) account.Account {
+				a, err := account.NewAccount(
+					account.ProviderM365,
+					account.M365Config{
+						M365: credentials.M365{
+							AzureClientID:     "Test",
+							AzureClientSecret: "without",
+						},
+						AzureTenantID: "data",
+					},
+				)
+				require.NoError(t, err, clues.ToCore(err))
+
+				return a
+			},
+		},
+	}
+
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			users, err := Users(ctx, test.acct(t), fault.New(true))
+			assert.Empty(t, users, "returned some users")
+			assert.NotNil(t, err)
+		})
+	}
+}
+
+func (suite *DiscoveryIntgSuite) TestSites() {
+	t := suite.T()
+
+	ctx, flush := tester.NewContext(t)
+	defer flush()
+
+	errs := fault.New(true)
+
+	sites, err := Sites(ctx, suite.acct, errs)
+	assert.NoError(t, err, clues.ToCore(err))
+
+	ferrs := errs.Errors()
+	assert.Nil(t, ferrs.Failure)
+	assert.Empty(t, ferrs.Recovered)
+	assert.NotEmpty(t, sites)
+}
+
+func (suite *DiscoveryIntgSuite) TestSites_InvalidCredentials() {
+	table := []struct {
+		name string
+		acct func(t *testing.T) account.Account
+	}{
+		{
+			name: "Invalid Credentials",
+			acct: func(t *testing.T) account.Account {
+				a, err := account.NewAccount(
+					account.ProviderM365,
+					account.M365Config{
+						M365: credentials.M365{
+							AzureClientID:     "Test",
+							AzureClientSecret: "without",
+						},
+						AzureTenantID: "data",
+					},
+				)
+				require.NoError(t, err, clues.ToCore(err))
+
+				return a
+			},
+		},
+		{
+			name: "Empty Credentials",
+			acct: func(t *testing.T) account.Account {
+				// intentionally swallowing the error here
+				a, _ := account.NewAccount(account.ProviderM365)
+				return a
+			},
+		},
+	}
+
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			sites, err := Sites(ctx, test.acct(t), fault.New(true))
+			assert.Empty(t, sites, "returned some sites")
+			assert.NotNil(t, err)
+		})
+	}
+}
+
+func (suite *DiscoveryIntgSuite) TestUserInfo() {
+	table := []struct {
+		name      string
+		user      string
+		expect    *api.UserInfo
+		expectErr require.ErrorAssertionFunc
+	}{
+		{
+			name: "standard test user",
+			user: tester.M365UserID(suite.T()),
+			expect: &api.UserInfo{
+				ServicesEnabled: map[path.ServiceType]struct{}{
+					path.ExchangeService: {},
+					path.OneDriveService: {},
+				},
+				Mailbox: api.MailboxInfo{
+					Purpose:              "user",
+					ErrGetMailBoxSetting: nil,
+				},
+			},
+			expectErr: require.NoError,
+		},
+		{
+			name: "user does not exist",
+			user: uuid.NewString(),
+			expect: &api.UserInfo{
+				ServicesEnabled: map[path.ServiceType]struct{}{},
+				Mailbox:         api.MailboxInfo{},
+			},
+			expectErr: require.Error,
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			result, err := GetUserInfo(ctx, suite.acct, test.user)
+			test.expectErr(t, err, clues.ToCore(err))
+
+			if err != nil {
+				return
+			}
+
+			assert.Equal(t, test.expect.ServicesEnabled, result.ServicesEnabled)
+		})
+	}
+}
+
+func (suite *DiscoveryIntgSuite) TestUserWithoutDrive() {
+	userID := tester.M365UserID(suite.T())
+
+	table := []struct {
+		name   string
+		user   string
+		expect *api.UserInfo
+	}{
+		{
+			name: "user without drive and exchange",
+			user: "a53c26f7-5100-4acb-a910-4d20960b2c19", // User: testevents@10rqc2.onmicrosoft.com
+			expect: &api.UserInfo{
+				ServicesEnabled: map[path.ServiceType]struct{}{},
+				Mailbox: api.MailboxInfo{
+					ErrGetMailBoxSetting: []error{api.ErrMailBoxSettingsNotFound},
+				},
+			},
+		},
+		{
+			name: "user with drive and exchange",
+			user: userID,
+			expect: &api.UserInfo{
+				ServicesEnabled: map[path.ServiceType]struct{}{
+					path.ExchangeService: {},
+					path.OneDriveService: {},
+				},
+				Mailbox: api.MailboxInfo{
+					Purpose:              "user",
+					ErrGetMailBoxSetting: []error{},
+				},
+			},
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			result, err := GetUserInfo(ctx, suite.acct, test.user)
+			require.NoError(t, err, clues.ToCore(err))
+			assert.Equal(t, test.expect.ServicesEnabled, result.ServicesEnabled)
+			assert.Equal(t, test.expect.Mailbox.ErrGetMailBoxSetting, result.Mailbox.ErrGetMailBoxSetting)
+			assert.Equal(t, test.expect.Mailbox.Purpose, result.Mailbox.Purpose)
 		})
 	}
 }
